api/requirements.txt
alembic
sqlalchemy
psycopg2
SQLAlchemy-Utils
Flask-SQLAlchemy
gunicorn
oslo.config
blinker
flask-cors
simplejson
sqlalchemy_jsonapi
